% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gif.R
\name{create_gif}
\alias{create_gif}
\title{Creates a gif file}
\usage{
create_gif(
  seeds,
  size,
  anchor_layout,
  hue_turn,
  color_scheme,
  color_subset_center,
  color_subset_width,
  movement,
  output_file
)
}
\arguments{
\item{seeds}{Values to set the random seeds in \code{gen_simplex}.}

\item{size}{Number of points.}

\item{anchor_layout}{Either "random", "spiral", or "diamond".}

\item{hue_turn}{Degree to rotate all hue values.}

\item{color_scheme}{Either "subset" or "full" for hue range.}

\item{color_subset_center}{Center hue value when using subset color_scheme}

\item{color_subset_width}{Width hue value when using subset color_scheme}

\item{movement}{Determines how the paths are animated. Options are "march"
and "glide". "march" moves small parts along the path. "glide" fills the path
entirely then shrinks it away.}

\item{output_file}{File to save the png.}
}
\value{
png file
}
\description{
Creates a gif file
}
