% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_png.R
\name{create_png}
\alias{create_png}
\title{Creates a png file}
\usage{
create_png(
  seeds,
  size,
  anchor_layout,
  hue_turn,
  color_scheme,
  color_subset_center,
  color_subset_width,
  alpha_taper,
  output_file
)
}
\arguments{
\item{seeds}{Values to set the random seeds in \code{gen_simplex}.}

\item{size}{Number of points.}

\item{anchor_layout}{Either "random", "spiral", or "diamond".}

\item{hue_turn}{Degree to rotate all hue values.}

\item{color_scheme}{Either "subset" or "full" for hue range.}

\item{color_subset_center}{Center hue value when using subset color_scheme}

\item{color_subset_width}{Width hue value when using subset color_scheme}

\item{alpha_taper}{Direction the paths taper. Options are "start", "end",
and "both". "start" fades away at the anchor points while "end" fades away in
the opposite direction. "both" fades in both direction.}

\item{output_file}{File to save the png.}
}
\value{
png file
}
\description{
Creates a png file
}
