% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_anchor_points}
\alias{get_anchor_points}
\title{Internal function to get the starting points and their attributes.}
\usage{
get_anchor_points(
  seeds,
  size,
  anchor_layout,
  hue_turn,
  color_scheme,
  color_subset_center,
  color_subset_width
)
}
\arguments{
\item{seeds}{Values to set the random seeds in \code{gen_simplex}.}

\item{size}{Number of points.}

\item{anchor_layout}{Either "random", "spiral", or "diamond".}

\item{hue_turn}{Degree to rotate all hue values.}

\item{color_scheme}{Either "subset" or "full" for hue range.}

\item{color_subset_center}{Center hue value when using subset color_scheme}

\item{color_subset_width}{Width hue value when using subset color_scheme}
}
\value{
Dataframe of points with attributes
}
\description{
Internal function to get the starting points and their attributes.
}
