% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_color}
\alias{get_color}
\title{Gets the hex code color based on angle and percentage.}
\usage{
get_color(angle, percentage)
}
\arguments{
\item{angle}{Angle in degrees [0, 360]. Closely matches hue in HCL.}

\item{percentage}{Percent [0, 100]. Closely matches chroma in HCL.}
}
\value{
Hex code for a color value.
}
\description{
\code{flowfieldsfigment} uses a base color scheme covering all hues.
}
\examples{
get_color(0, 80)
get_color(90, 50)
}
