% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_color_subset}
\alias{get_color_subset}
\title{Gets the hex code color from get_color augmented by center and width.}
\usage{
get_color_subset(center, width, angle, percentage)
}
\arguments{
\item{center}{Center hue angle [0, 360] for the color scheme.}

\item{width}{Width in angles from center to edge for hue values.}

\item{angle}{Angle in degrees [0, 360]. Closely matches hue in HCL.}

\item{percentage}{Percent [0, 100]. Closely matches chroma in HCL.}
}
\value{
Hex code for a color value.
}
\description{
This function converts the color scheme to use a subset of hues.
\code{get_color_subset} maps the the angle to be \code{center} when 0 and 180
and stretch to \code{center - width} at 90 to \code{center + width} as 270.
}
\examples{
get_color_subset(90, 30, 90, 100)
get_color_subset(180, 90, 270, 80)
}
