% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_paths}
\alias{get_paths}
\title{Internal function to get the points' full paths.}
\usage{
get_paths(points, seeds)
}
\arguments{
\item{points}{Current position of the points.}

\item{seeds}{Values to set the random seeds in \code{gen_simplex}.}
}
\value{
Dataframe of paths.
}
\description{
This starts points along their paths. Paths are shorten when either the
points hit a plateau (vectors are 0 for both x and y) or when filtered by
percentage.
}
