% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_point_paths}
\alias{get_point_paths}
\title{Internal function to handle edge case of when points start on a plateau.}
\usage{
get_point_paths(points, paths)
}
\arguments{
\item{points}{Current position of the points.}

\item{paths}{Paths the points take.}
}
\value{
Dataframe of points with attributes for time 0.
}
\description{
Internal function to handle edge case of when points start on a plateau.
}
