% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_vectors}
\alias{get_vectors}
\title{Internal function to get directional vectors.}
\usage{
get_vectors(points, seeds)
}
\arguments{
\item{points}{Current position of the points.}

\item{seeds}{Values to set the random seeds in \code{gen_simplex}.}
}
\value{
The direction vectors the points should move.
}
\description{
This function uses the \code{ambient} package to generate noise for the
points to follow.
}
