% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{move_points}
\alias{move_points}
\title{Internal function to move points by one time period.}
\usage{
move_points(points, seeds)
}
\arguments{
\item{points}{Current position of the points.}

\item{seeds}{Values to set the random seeds in \code{gen_simplex}.}
}
\value{
The next steps for points.
}
\description{
This function uses \code{get_vectors} to determine where
the points should go.
}
